separated class string logic to a separate class to make OpBase simpler. Also added unit test-reports and png files in main dir to .gitignore
